Uses apt recipe for update instead of shelling out for it manually
The presence of an `execute[apt-get update]` resource has the potential to conflict with other recipes which have a resource of an identical name - including the default apt recipe! By including the default apt recipe, we'll trigger an apt-get update in a reusable manner, and hopefully this method can also eliminate redundantly `apt-get update`ing multiple times in a single converge.